Created simple timer
Made a simple countdown timer, you can change timeRemaining variable to the desired amount.

Uses text mesh pro, hopefully not an issue. If it is I can change it.

To use, drag the Timer.cs file onto the text and then drag the text into the "Time Text" box in the script section and it should work.